Added block and chain tip related event notification mechanism for model/chain.
