+++ model/chain/const.go
@@ -1,53 +1,5 @@
 package chain
 
-import "fmt"
-
-// NotificationType represents the type of a notification message.
-type NotificationType int
-
-// Constants for the type of a notification message.
-const (
-	// NTBlockAccepted indicates the associated block was accepted into
-	// the block chain.  Note that this does not necessarily mean it was
-	// added to the main chain.  For that, use NTBlockConnected.
-	NTBlockAccepted NotificationType = iota
-
-	// NTBlockConnected indicates the associated block was connected to the
-	// main chain.
-	NTBlockConnected
-
-	// NTBlockDisconnected indicates the associated block was disconnected
-	// from the main chain.
-	NTBlockDisconnected
-)
-
-// notificationTypeStrings is a map of notification types back to their constant
-// names for pretty printing.
-var notificationTypeStrings = map[NotificationType]string{
-	NTBlockAccepted:     "NTBlockAccepted",
-	NTBlockConnected:    "NTBlockConnected",
-	NTBlockDisconnected: "NTBlockDisconnected",
-}
-
-// String returns the NotificationType in human-readable form.
-func (n NotificationType) String() string {
-	if s, ok := notificationTypeStrings[n]; ok {
-		return s
-	}
-	return fmt.Sprintf("Unknown Notification Type (%d)", int(n))
-}
-
-// Notification defines notification that is sent to the caller via the callback
-// function provided during the call to New and consists of a notification type
-// as well as associated data that depends on the type as follows:
-// 	- NTBlockAccepted:     *btcutil.Block
-// 	- NTBlockConnected:    *btcutil.Block
-// 	- NTBlockDisconnected: *btcutil.Block
-type Notification struct {
-	Type NotificationType
-	Data interface{}
-}
-
 // BehaviorFlags is a bitmask defining tweaks to the normal behavior when
 // performing chain processing and consensus rules checks.
 type BehaviorFlags uint32
